Refactor Heartbeat class into its own module and replace heartbeat callback in stm_unique_id.py with instance of Heartbeat class.
